requirements.txt
 streamlit 
 ultralytics 
 reportlab 
 pillow